# Making Verbnet feature more granular