--- a/tools/parse/src/main/resources/featuretaggers.config
+++ b/tools/parse/src/main/resources/featuretaggers.config
@@ -1,0 +1,5 @@
+verbnet {
+  group: "org.allenai.nlp.resources"
+  name: "verbnet-3.2"
+  version: 1
+}