--- a/tools/parse/src/main/resources/polyparser.config
+++ b/tools/parse/src/main/resources/polyparser.config
@@ -1,5 +1,0 @@
-verbnet {
-  group: "org.allenai.nlp.resources"
-  name: "verbnet-3.2"
-  version: 1
-}