--- a/tools/parse/src/main/scala/org/allenai/nlpstack/parse/poly/core/SentenceTransform.scala
+++ b/tools/parse/src/main/scala/org/allenai/nlpstack/parse/poly/core/SentenceTransform.scala
@@ -158,12 +158,7 @@ case class VerbnetTagger(verbnet: Verbnet) extends SentenceTransform {
       val tokLemmaLC = tok.getDeterministicProperty('factorieLemma).name.toLowerCase
       val tokVerbnetPrimaryFrames = verbnet.getVerbnetFramePrimaryNames(tokLemmaLC)
       val tokVerbnetSecondaryFrames = verbnet.getVerbnetFrameSecondaryNames(tokLemmaLC)
-      tok.updateProperties(
-        Map(
-          'verbnetPrimaryFrames -> tokVerbnetPrimaryFrames,
-          'verbnetSecondaryFrames -> tokVerbnetSecondaryFrames
-        )
-      )
+      tok.updateProperties(Map('verbnetPrimaryFrames -> tokVerbnetPrimaryFrames))
     })
   }
 }