--- a/tools/parse/src/main/scala/org/allenai/nlpstack/parse/poly/polyparser/DependencyParsingTransitionSystem.scala
+++ b/tools/parse/src/main/scala/org/allenai/nlpstack/parse/poly/polyparser/DependencyParsingTransitionSystem.scala
@@ -64,6 +64,8 @@ abstract class DependencyParsingTransitionSystem(
       TokenPropertyFeature('factorieCpos),
       TokenPropertyFeature('factoriePos),
       TokenPropertyFeature('brown0),
+      TokenPropertyFeature('verbnetPrimaryFrames),
+      TokenPropertyFeature('verbnetSecondaryFrames),
       KeywordFeature(DependencyParsingTransitionSystem.keywords)
     ))
     tokenFeatureTagger.tag(overriddenSentence)