--- a/tools/parse/src/test/scala/org/allenai/nlpstack/parse/poly/fsm/TrainingVectorSourceSpec.scala
+++ b/tools/parse/src/test/scala/org/allenai/nlpstack/parse/poly/fsm/TrainingVectorSourceSpec.scala
@@ -2,9 +2,6 @@ package org.allenai.nlpstack.parse.poly.fsm
 
 import org.allenai.common.testkit.UnitSpec
 
-
 class TrainingVectorSourceSpec extends UnitSpec {
-
-
 
 }