@mhrmm : Changes as discussed. Some improvement in scores.
Note that based on the table at the top of this page: http://verbs.colorado.edu/~mpalmer/projects/verbnet.html#frames, you had suggested using the Frame Name and the Frame Syntax. For the e.g. in that table (`hit-18.1`), this is what the frame xml snippet looks like:

```
        <FRAME>
            <DESCRIPTION descriptionNumber="0.2" primary="NP V NP" secondary="Basic Transitive" xtag="0.2"/>
            <EXAMPLES>
                <EXAMPLE>Paula hit the ball.</EXAMPLE>
            </EXAMPLES>
            <SYNTAX>
                <NP value="Agent">
                    <SYNRESTRS/>
                </NP>
                <VERB/>
                <NP value="Patient">
                    <SYNRESTRS/>
                </NP>
            </SYNTAX>
...
```

`Basic Transitive` is the secondary name, and `NP V NP` is the primary frame name, which represents the syntax `Agent V Patient`. So I have used the primary frame names and the secondary frame names as the features. Since a class can have multiple frames, there can be multiple values (a set) coming from the same class for a given verb.

The `featuretaggers.config` (rename of the earlier `polyparser.config`) can be used to pass any config info for any other features, for e.g., the Brown clusters, which is not in the datastore yet.

Scores

Baseline

```
UAS: 9084 / 10427 = 87.12%
UAS (no punc): 8019 / 9130 = 87.83%
LAS: 8663 / 10427 = 83.08%
LAS (no punc): 7598 / 9130 = 83.22%
Path Accuracy (labeled, full): 7371 / 10427 = 70.69%
Path Accuracy (unlabeled, full): 7903 / 10427 = 75.79%
Path Accuracy (labeled, ignorePunc): 6347 / 9130 = 69.52%
Path Accuracy (unlabeled, ignorePunc): 6875 / 9130 = 75.30%
Parsed 1000 sentences in 16.3 seconds, an average of 61.4 sentences per second.
```


With Verbnet

```
UAS: 9142 / 10427 = 87.68%
UAS (no punc): 8059 / 9130 = 88.27%
LAS: 8741 / 10427 = 83.83%
LAS (no punc): 7659 / 9130 = 83.89%
Path Accuracy (labeled, full): 7515 / 10427 = 72.07%
Path Accuracy (unlabeled, full): 8007 / 10427 = 76.79%
Path Accuracy (labeled, ignorePunc): 6465 / 9130 = 70.81%
Path Accuracy (unlabeled, ignorePunc): 6954 / 9130 = 76.17%
Parsed 1000 sentences in 19.4 seconds, an average of 51.4 sentences per second.
```


With Brown Clusters

```
UAS: 9154 / 10427 = 87.79%
UAS (no punc): 8084 / 9130 = 88.54%
LAS: 8780 / 10427 = 84.20%
LAS (no punc): 7710 / 9130 = 84.45%
Path Accuracy (labeled, full): 7505 / 10427 = 71.98%
Path Accuracy (unlabeled, full): 8027 / 10427 = 76.98%
Path Accuracy (labeled, ignorePunc): 6462 / 9130 = 70.78%
Path Accuracy (unlabeled, ignorePunc): 6980 / 9130 = 76.45%
Parsed 1000 sentences in 20.0 seconds, an average of 49.9 sentences per second.
```

With Verbnet + Brown Clusters

```
UAS: 9215 / 10427 = 88.38%
UAS (no punc): 8114 / 9130 = 88.87%
LAS: 8835 / 10427 = 84.73%
LAS (no punc): 7736 / 9130 = 84.73%
Path Accuracy (labeled, full): 7570 / 10427 = 72.60%
Path Accuracy (unlabeled, full): 8090 / 10427 = 77.59%
Path Accuracy (labeled, ignorePunc): 6510 / 9130 = 71.30%
Path Accuracy (unlabeled, ignorePunc): 7026 / 9130 = 76.96%
Parsed 1000 sentences in 24.4 seconds, an average of 41.0 sentences per second.
```
Side note: Training takes a lot longer with the Verbnet features, as recorded in the final timing line.