@sbhaktha 2015-03-18
@mhrmm : Check it out now. Test passes. Embarrassing bug fixed in this commit: https://github.com/sbhaktha/nlpstack/commit/c4f32c1a09fc2deeccd7103700c282706d763703 .

Scores:

Primary Frames Only

```
[info] UAS: 9138 / 10427 = 87.64%
[info] UAS (no punc): 8061 / 9130 = 88.29%
[info] LAS: 8737 / 10427 = 83.79%
[info] LAS (no punc): 7660 / 9130 = 83.90%
[info] Path Accuracy (labeled, full): 7465 / 10427 = 71.59%
[info] Path Accuracy (unlabeled, full): 7961 / 10427 = 76.35%
[info] Path Accuracy (labeled, ignorePunc): 6430 / 9130 = 70.43%
[info] Path Accuracy (unlabeled, ignorePunc): 6921 / 9130 = 75.81%
[info] Parsed 1000 sentences in 20.4 seconds, an average of 49.0 sentences per second.
```

Secondary Frames Only

```
[info] UAS: 9142 / 10427 = 87.68%
[info] UAS (no punc): 8059 / 9130 = 88.27%
[info] LAS: 8741 / 10427 = 83.83%
[info] LAS (no punc): 7659 / 9130 = 83.89%
[info] Path Accuracy (labeled, full): 7515 / 10427 = 72.07%
[info] Path Accuracy (unlabeled, full): 8007 / 10427 = 76.79%
[info] Path Accuracy (labeled, ignorePunc): 6465 / 9130 = 70.81%
[info] Path Accuracy (unlabeled, ignorePunc): 6954 / 9130 = 76.17%
[info] Parsed 1000 sentences in 19.4 seconds, an average of 51.4 sentences per second.
```

Primary + Secondary

```
[info] UAS: 9119 / 10427 = 87.46%
[info] UAS (no punc): 8044 / 9130 = 88.11%
[info] LAS: 8708 / 10427 = 83.51%
[info] LAS (no punc): 7633 / 9130 = 83.60%
[info] Path Accuracy (labeled, full): 7443 / 10427 = 71.38%
[info] Path Accuracy (unlabeled, full): 7944 / 10427 = 76.19%
[info] Path Accuracy (labeled, ignorePunc): 6404 / 9130 = 70.14%
[info] Path Accuracy (unlabeled, ignorePunc): 6903 / 9130 = 75.61%
[info] Parsed 1000 sentences in 21.9 seconds, an average of 45.7 sentences per second.
```

If `verbnet` feature requested (per training cmd line), by default both features are included currently. But `verbnet` is off by default anyway. If you think I should change it so only Secondary frame features are included (based on max scores above) lmk.