@sbhaktha 2015-03-19
@mhrmm : Check out latest commit, addresses above points. 

Also dumped out some Verbnet data and stats today. Here are some histograms, though I am not sure this is what you had in mind. I can iterate with you on it, but here are some:

https://drive.google.com/drive/#folders/0B971Sd5f2y4HfmkyaHdnZy01RWtBR3dISlo3bmJ1bE1YSndSYUZ2SVdhYmdiYjRWQUJNams

`primaryFrameVerbStats, secondaryFrameVerbStats : `
x-axis : verb count bins
y-axis : no. of primary/secondary frames respectively
— can be read as, for e.g., “most primary frames have 1-5 different verbs associated with them”.


`verbPrimaryFrameStats, verbSecondaryFrameStats:`
x-axis : primary/secondary frame count bins respectively
y-axis : no. of verbs
— can be read as, for e.g., “most verbs occupy 3-4 different primary frames”.