@sbhaktha 2015-03-19
I also have plots of actual frames to no. of verbs, and the other way round to get an idea of the distribution. I'll put those up tonight as well.