@sbhaktha 2015-03-20
@mhrmm Using only primary frames. Removed role suffixes such as `.Theme` from syntactic constituents.
Fixes based on your two above comments.